requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# Core Application Dependencies  
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
apify-client==1.7.1
aiohttp==3.9.1
asyncio-throttle==1.0.2
fake-useragent==1.4.0
lxml==4.9.3
openai>=1.0.0

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
structlog==23.2.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0